Fix bug in dual LoTs
In previous version, dual LoTs were considered equivalent from the point of view of learning. However, they are not (see code for more exhaustive explanation). In new code all functionally complete LoTs are considered, except when having a negation is useless in which case the LoT *with* the negation is not considered.